feat: Create new file for Lambda Group server
