+++ src/example/multi.ts
@@ -1,6 +1,6 @@
 import { Context } from 'aws-lambda';
-import { LocalLambdaGroupConfig, LambdaConfig, LocalLambdaGroup } from '../local.lambda';
 import { LambdaResponse, RequestEvent } from '../types';
+import { LambdaConfig, LocalLambdaGroupConfig, LocalLambdaGroup } from '../lambda.group';
 
 // handler is a function that takes in an event and context and returns a response
 const handler = async (req: RequestEvent, context: Context): Promise<LambdaResponse> => ({ statusCode: 200, body: `Hello World !!! My userId is ${req.pathParameters?.id}\n My JWT is ${JSON.stringify(req.requestContext.authorizer.lambda.jwt)}. My queryStringParameter is ${JSON.stringify(req.queryStringParameters)} ` });
@@ -51,5 +51,7 @@ const multiConfig: LocalLambdaGroupConfig = {
   port: 8008, // optional, default to 8000
   defaultPath: '/api/v1', // optional, default to '/'
 };
+
+// visit http://localhost:8008/user/1234567890 to see the response
 const localLambdaGroup = new LocalLambdaGroup(multiConfig);
 localLambdaGroup.run();
