+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './local.lambda';
 export * from './types/index';
+export * from './lambda.group';
