+++ src/local.lambda.ts
@@ -4,7 +4,8 @@ import HTTPMethod from 'http-method-enum';
 import { LambdaHandler, RequestEvent } from './types';
 import express, { Request, Response } from 'express';
 import { flattenArraysInJSON, cloneDeep } from './utils';
-const DefaultPort = 8000;
+import { LambdaConfig } from './lambda.group';
+export const DefaultPort = 8000;
 
 // binary upload content-type headers
 const defaultBinaryContentTypeHeaders = [
@@ -16,7 +17,7 @@ const defaultBinaryContentTypeHeaders = [
   'application/zip',
 ];
 
-const DefaultPathParamsPattern = '/';
+export const DefaultPathParamsPattern = '/';
 
 export class LocalLambda {
   handler: LambdaHandler;
@@ -105,47 +106,7 @@ export class LocalLambda {
 
 }
 
-
-export interface LambdaConfig {
-  handler: LambdaHandler;
-  context?: Context;
-  enableCORS?: boolean;
-  // default binary content-type headers or override
-  binaryContentTypesOverride?: string[];
-  pathParamsPattern ?: string;
-  requestContext?: Record<string, any>;
-}
-
 // extend the LambdaConfig to add port
 export interface LocalLambdaConfig extends LambdaConfig {
   port?: number;
-}
-
-export interface LocalLambdaGroupConfig {
-  lambdas: LambdaConfig[];
-  port?: number;
-  defaultPath?: string;
-}
-
-export class LocalLambdaGroup {
-  lambdas: LambdaConfig[] = [];
-  app: express.Application;
-  port: number;
-  defaultPath: string;
-
-  constructor(config: LocalLambdaGroupConfig) {
-    this.lambdas = config.lambdas;
-    this.app = express();
-    this.port = config.port ?? DefaultPort;
-    this.defaultPath = config.defaultPath ?? DefaultPathParamsPattern;
-  }
-
-  run(): void {
-    this.lambdas.forEach(lambda => {
-      const localLambda = new LocalLambda(lambda, this.app, this.defaultPath);
-      localLambda.createServer();
-      this.app = localLambda.app;
-    });
-    this.app.listen(this.port, () => console.info(`🚀  Lambda Group Server ready at http://localhost:${this.port} at '${new Date().toLocaleString()}'`));
-  }
 }
